+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Fix a bug related to incorrect start of celery ([Issue](https://github.com/orgs/amosproj/projects/79/views/2?pane=issue&itemId=115359765&issue=amosproj%7Camos2025ss01-embark-orchestration-framework%7C79))
 - Fix a bug related to collecting worker dependency information
+- Fix a bug related to collecting worker external repo and emba version information
 
 ### ADDED
 
@@ -13,6 +14,7 @@
 - Collect dependency version info for workers ([Issue](https://github.com/orgs/amosproj/projects/79/views/2?pane=issue&itemId=114773112&issue=amosproj%7Camos2025ss01-embark-orchestration-framework%7C73))
 - Settings App to EMBArk ([Issue](https://github.com/orgs/amosproj/projects/79/views/2?pane=issue&itemId=114643627&issue=amosproj%7Camos2025ss01-embark-orchestration-framework%7C66))
 - Settings helper `workers_enabled` check
+- Celery for reset ([Issue](https://github.com/orgs/amosproj/projects/79/views/2?pane=issue&itemId=115510818&issue=amosproj%7Camos2025ss01-embark-orchestration-framework%7C81))
 
 ### CHANGED